+++ src/components/trip-events-list.js
@@ -2,13 +2,10 @@ import {createTripEventTemplate} from "../components/trip-event.js";
 
 export const createTripEventsListTemplate = (events) => {
   const getEvents = () => {
-    const eventsList = [];
-    events.forEach((it) => {
-      eventsList.push(createTripEventTemplate(it));
-    }
-    );
-
-    return eventsList.join(`\n`);
+    return events.map((it) => {
+      return createTripEventTemplate(it);
+    })
+    .join(`\n`);
   };
   return (
     `<ul class="trip-events__list">
